Refactored the authentication
Now Rpc instances can get the authentication later.

Will work on the database issue later


Former-commit-id: 3139b985b6302fcc1a46121d2397c7090c59212d
