+++ CourseSketchProjects/ServerAbstractor/src/main/java/coursesketch/server/interfaces/AbstractServerWebSocketHandler.java
@@ -2,6 +2,7 @@ package coursesketch.server.interfaces;
 
 import com.google.protobuf.InvalidProtocolBufferException;
 import coursesketch.database.interfaces.AbstractCourseSketchDatabaseReader;
+import database.DatabaseAccessException;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import protobuf.srl.request.Message;
@@ -332,7 +333,11 @@ public abstract class AbstractServerWebSocketHandler {
      */
     public final void initialize() {
         databaseReader = createDatabaseReader(this.serverInfo);
-        startDatabase();
+        try {
+            startDatabase();
+        } catch (DatabaseAccessException e) {
+            LOG.error("An error was created starting the database for the server", e);
+        }
         onInitialize();
     }
 
@@ -346,8 +351,10 @@ public abstract class AbstractServerWebSocketHandler {
 
     /**
      * Starts the database if it exists.
+     *
+     * @throws DatabaseAccessException thrown if the database is unable to start.
      */
-    protected final void startDatabase() {
+    protected final void startDatabase() throws DatabaseAccessException {
         final AbstractCourseSketchDatabaseReader reader = getDatabaseReader();
         if (reader != null) {
             reader.startDatabase();
